+++ cloudplatform-webportal-sdk-billing/src/main/java/com/letv/portal/model/subscription/SubscriptionDetail.java
@@ -1,5 +1,6 @@
 package com.letv.portal.model.subscription;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import org.apache.ibatis.type.Alias;
@@ -27,13 +28,13 @@ public class SubscriptionDetail extends BaseModel{
 	private Date endTime;//结束时间
 	private boolean valid;//是否有效：false-无效，true-有效
 	private String descn;//描述
-	private Double price;//价格-用于记录各规格价格
+	private BigDecimal price;//价格-用于记录各规格价格
 	
 	
-	public Double getPrice() {
+	public BigDecimal getPrice() {
 		return price;
 	}
-	public void setPrice(Double price) {
+	public void setPrice(BigDecimal price) {
 		this.price = price;
 	}
 	public String getDescn() {
